+++ src/LanguageServer.cpp
@@ -258,13 +258,19 @@ void LanguageServer::onRequest(const id_type& id, const std::string& method, std
         auto workspace = findWorkspace(params.textDocument.uri);
         response = workspace->bytecode(params);
     }
-
     else if (method == "luau-lsp/compilerRemarks")
     {
         ASSERT_PARAMS(baseParams, "luau-lsp/compilerRemarks")
         auto params = baseParams->get<lsp::CompilerRemarksParams>();
         auto workspace = findWorkspace(params.textDocument.uri);
         response = workspace->compilerRemarks(params);
+    }
+    else if (method == "luau-lsp/codegen")
+    {
+        ASSERT_PARAMS(baseParams, "luau-lsp/codeGen")
+        auto params = baseParams->get<lsp::CodegenParams>();
+        auto workspace = findWorkspace(params.textDocument.uri);
+        response = workspace->codeGen(params);
     }
     else
     {